Added warning for buttons that are already mapped to something else
Button text will now turn red if the button is already assigned to a separate action. Also fixed some spelling mistakes
