+++ Monster Mash/Monster Mash/Assets/InputRemapper.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b27f81ff0a703384f906e0afe46777ca
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
