+++ Monster Mash/Monster Mash/Assets/UI/InputRemaper.prefab
@@ -929,7 +929,7 @@ GameObject:
   - component: {fileID: 9028997243404824085}
   - component: {fileID: 9028997243404824084}
   m_Layer: 0
-  m_Name: RemapingUI
+  m_Name: RemappingUI
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -1007,7 +1007,7 @@ GameObject:
   - component: {fileID: 9028997243670024169}
   - component: {fileID: 9028997243670024168}
   m_Layer: 0
-  m_Name: InputRemapingHeaderText
+  m_Name: InputRemappingHeaderText
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -2159,8 +2159,3 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: b27f81ff0a703384f906e0afe46777ca, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  allControlItems:
-  - {fileID: 9028997244117319017}
-  - {fileID: 9028997244413465722}
-  - {fileID: 9028997243084044401}
-  - {fileID: 9028997243305268408}
